BLE L2CAP: On Android, ensure only a single L2CAP socket is open.
For L2CAP there is no flush() call so we wait a bit before closing the
socket otherwise data is lost. In particular with ZKP this is a big
problem because it can take 15 seconds to transfer the 400 kB proof.

Recent versions of Android seem to have a problem if two L2CAP sockets
are open at the same time so add code to ensure that we close the
deferred socket sooner in case we wish to create a new one. This
ensures that only a single L2CAP socket is open at any given time.

Test: Manually tested several successful L2CAP presentations in sequence.
Signed-off-by: David Zeuthen <zeuthen@google.com>