Observe Mode support.
This adds support for [ECP and Observe Mode](https://github.com/kormax/apple-enhanced-contactless-polling)
both on the wallet and reader side.

On the wallet side this is currently specific to samples/testapp as it
involves an HostApduService and it'll be a bit more work to generalize
into library form.

On the reader side a new `NfcScanOptions` class is introduced which
can be passed to NfcTagReader.scan() and `NfcTagReader.scanMdocReader()`
which allows the app to pass the payload to be inserted into the NFC
polling loop.

This works on Android 16 and later on supported Android devices. It
has been tested with Pixel Pro 8 and Pixel Pro 10 XL. There is no iOS
support for this at this point.

Fixes #1078 and #1079.

Also rename some source directories in testapp from
`com/android/identity/testapp` to `org/multipaz/testapp`.

Test: Manually tested testapp on both Android and iOS.
Signed-off-by: David Zeuthen <zeuthen@google.com>