NFC: Fix issues with errors when dialog is hidden.
For the case where we're scanning a holder's Android device with two
wallets and there's no wallet role owner (or the wallet role owner
isn't registering the NDEF AID), Android shows a disambiguation dialog
which results in transceive() failing when the two devices are no
longer in the field.

The application - e.g. Multipaz Identity Reader - should be able to
catch the exception and then restart NFC scanning as a NFC tag
reader. However in this situation the following things happen

- Despite the prompt being invisible, it shows the error
- The app must wait for linger duration until it can start reading again.

This PR fixes these two problems.

Test: Manually tested with Multipaz Identity Reader.
Signed-off-by: David Zeuthen <zeuthen@google.com>