Fix problems with running testapp on iOS.
First of all, updating to latest androidx-sqlite version (2.6.2) made
things stop compiling for iOS with undefined symbols. Downgrade to the
version (2.5.2) to fix this.

Also to make things work on iOS also update to the latest
jetbrains-navigation release. This includes a non-trivial change to
testapp and how navigation works because strings are no longer used
for routes.

Test: Manually tested on iOS and Android.
Signed-off-by: David Zeuthen <zeuthen@google.com>