ConsentScreen and SD-JWT fixes.
Looks like we're not including the `x5c` claim when creating an SD-JWT
and the issuer signing key has a X.509 certificate chain. Fix this.

Also bunch of fixes for the ConsentScreen in testapp to make it work
with the latest update to the Compose libraries which fixes some bugs
the existing code was relying on.

Also temporarily disable running unit tests in the Android emulator
since it current breaks our GitHub runner due to insufficient disk
space.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>